requirements.txt
asgiref==3.7.2
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
defusedxml==0.7.1
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
easy-thumbnails==2.8.5
idna==3.4
inflate64==0.3.1
lxml==4.9.3
multivolumefile==0.2.3
mysqlclient==2.2.0
Pillow==10.0.1
psutil==5.9.6
pybcj==1.0.1
pycryptodomex==3.19.0
pyppmd==1.0.0
pytz==2023.3.post1
pyzstd==0.15.9
requests==2.31.0
sqlparse==0.4.4
texttable==1.7.0
typing_extensions==4.8.0
urllib3==1.26.17
xmltodict==0.13.0
